usb: dwc3: Drive a pulse on DP on CDP detection
Some USB host PCs switch the downstream port mode from
CDP to SDP if the device does not pull DP up within certain
duration (~2s).
This limits the current drawn by the device to 500mA(or 900mA)
as opposed to the possible 1.5A resulting in slow charging of
the battery.
Fix this by driving a DP pulse when there is a connect
notification from PMIC and detected the charger type as CDP.

Change-Id: Ia04e3c3f2dc9ad96781d2aa91f8842b6aa8dc0da
Signed-off-by: Udipto Goswami <ugoswami@codeaurora.org>